Move Key from reporting.utils to (new) reporting.key
